chore(lint): lint 설정 변경

+++ src/comment/comment.service.ts
@@ -1,48 +1,60 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
+import { Repository, type FindOneOptions } from 'typeorm';
 import { PostService } from 'src/post/post.service';
 import { Comment } from './entities/comment.entity';
 import { User } from 'src/users/entities/user.entity';
-import { CreateCommentDto } from './dto/create-comment.dto';
-import { CreateReplyDto } from './dto/create-reply.dto';
-import { UpdateCommentDto } from './dto/update-comment.dto';
-import { FindOneOptions } from 'typeorm';
-import { NullableType } from 'src/utils/types/nullable.type';
+import { type CreateCommentDto } from './dto/create-comment.dto';
+import { type CreateReplyDto } from './dto/create-reply.dto';
+import { type UpdateCommentDto } from './dto/update-comment.dto';
+import { type NullableType } from 'src/utils/types/nullable.type';
 
 @Injectable()
 export class CommentService {
     constructor(
         @InjectRepository(Comment)
-        private commentRepository: Repository<Comment>,
+        private readonly commentRepository: Repository<Comment>,
         @InjectRepository(User)
-        private userRepository: Repository<User>,
-        private postService: PostService
-    ) { }
+        private readonly userRepository: Repository<User>,
+        private readonly postService: PostService,
+    ) {}
 
-    async create(createCommentDto: CreateCommentDto, user: User): Promise<Comment> {
+    async create(
+        createCommentDto: CreateCommentDto,
+        user: User,
+    ): Promise<Comment> {
         const { postId } = createCommentDto;
         const post = await this.postService.findOne({
             where: {
-                id: postId
-            }
+                id: postId,
+            },
         });
 
-        return this.commentRepository.save(
+        if (!post) {
+            throw new NotFoundException('게시물이 존재하지 않습니다.');
+        }
+
+        const comment = await this.commentRepository.save(
             this.commentRepository.create({
                 ...createCommentDto,
                 post, // 연결된 게시물
-                user
-            })
+                user,
+            }),
         );
+
+        return comment;
     }
 
-    async createReply(createReplyDto: CreateReplyDto, parentCommentId: number, user: User): Promise<Comment> {
+    async createReply(
+        createReplyDto: CreateReplyDto,
+        parentCommentId: number,
+        user: User,
+    ): Promise<Comment> {
         const { replyToId } = createReplyDto;
         const replyTo = await this.userRepository.findOne({
             where: {
-                id: replyToId
-            }
+                id: replyToId,
+            },
         });
         if (!replyTo) {
             throw new NotFoundException('존재하지 않는 사용자입니다.');
@@ -50,30 +62,35 @@ export class CommentService {
 
         const parentComment = await this.findOne({
             where: {
-                id: parentCommentId
+                id: parentCommentId,
             },
-            relations: ['post']
+            relations: ['post'],
         });
-        if (!parentComment) throw new NotFoundException('존재하지 않는 댓글입니다.');
+        if (!parentComment) {
+            throw new NotFoundException('존재하지 않는 댓글입니다.');
+        }
+
         const post = parentComment.post;
 
-        return this.commentRepository.save(
+        return await this.commentRepository.save(
             this.commentRepository.create({
                 ...createReplyDto,
                 user,
                 replyTo,
                 post,
-                parentComment
-            })
+                parentComment,
+            }),
         );
     }
 
-    async findOne(findOptions: FindOneOptions<Comment>): Promise<NullableType<Comment>> {
-        return this.commentRepository.findOne(findOptions);
+    async findOne(
+        findOptions: FindOneOptions<Comment>,
+    ): Promise<NullableType<Comment>> {
+        return await this.commentRepository.findOne(findOptions);
     }
 
-    async findMany(postId: number): Promise<NullableType<Comment[]>> {
-        const comments = [];
+    async findMany(postId: number): Promise<Comment[]> {
+        const comments: Comment[] = [];
         const rootComments = await this.commentRepository
             .createQueryBuilder('comment')
             .leftJoinAndSelect('comment.user', 'user')
@@ -87,7 +104,7 @@ export class CommentService {
                 'user.id',
                 'user.nickname',
                 'user.userId',
-                'user.profileImage'
+                'user.profileImage',
             ])
             .getMany();
 
@@ -97,7 +114,9 @@ export class CommentService {
                 .createQueryBuilder('comment')
                 .leftJoinAndSelect('comment.user', 'user')
                 .leftJoinAndSelect('comment.replyTo', 'replyTo')
-                .where('comment.parentCommentId = :parentId', { parentId: rootComment.id })
+                .where('comment.parentCommentId = :parentId', {
+                    parentId: rootComment.id,
+                })
                 .andWhere('comment.postId = :postId', { postId }) // 추가된 부분
                 .orderBy('comment.createdAt', 'ASC')
                 .select([
@@ -111,7 +130,7 @@ export class CommentService {
                     'user.userId',
                     'replyTo.id',
                     'replyTo.nickname',
-                    'replyTo.userId'
+                    'replyTo.userId',
                 ])
                 .getMany();
             rootComment.replies = replies;
@@ -121,24 +140,34 @@ export class CommentService {
         return comments;
     }
 
-    async update(id: number, updateCommentDto: UpdateCommentDto): Promise<Comment> {
+    async update(
+        id: number,
+        updateCommentDto: UpdateCommentDto,
+    ): Promise<Comment> {
         const { content } = updateCommentDto;
         const comment = await this.findOne({
             where: {
-                id
-            }
+                id,
+            },
         });
+        if (!comment) {
+            throw new NotFoundException('댓글이 존재하지 않습니다.');
+        }
         comment.content = content;
 
-        return this.commentRepository.save(comment);
+        return await this.commentRepository.save(comment);
     }
 
     async delete(id: number): Promise<void> {
         const comment = await this.findOne({
             where: {
-                id
-            }
+                id,
+            },
         });
+        if (!comment) {
+            throw new NotFoundException('댓글이 존재하지 않습니다.');
+        }
+
         await this.commentRepository.remove(comment);
     }
 }
